+++ lib/src/models/typedef.dart
@@ -23,20 +23,6 @@ typedef EventDayViewItemBuilder<T extends Object> = Widget Function(
   DayEvent<T> event,
 );
 
-typedef CategoryBackgroundTimeRowBuilder = Widget Function(
-  BuildContext context,
-  BoxConstraints constraints,
-  DateTime rowTime,
-  bool isOdd,
-);
-typedef CategoryBackgroundTimeTileBuilder = Widget Function(
-  BuildContext context,
-  BoxConstraints constraints,
-  DateTime rowTime,
-  EventCategory category,
-  bool isOddRow,
-);
-
 typedef OnTimeTap = Function(DateTime time);
 
 typedef CategoryDayViewEventBuilder<T extends Object> = Widget Function(
@@ -45,35 +31,12 @@ typedef CategoryDayViewEventBuilder<T extends Object> = Widget Function(
   DateTime time,
   CategorizedDayEvent<T> event,
 );
-typedef CategoryDayViewRowBuilder<T extends Object> = Widget Function(
-  List<EventCategory> category,
-  List<CategorizedDayEvent<T>> events,
-  DateTime time,
-);
 typedef CategoryDayViewTileTap<T extends Object> = Function(
   EventCategory category,
   DateTime time,
-);
-
-/// To build the controller bar on the top of the day view
-///
-/// [goToPreviousTab] to animate to previous tabs
-/// [goToNextTab] to animate to next tabs
-typedef CategoryDayViewControlBarBuilder = Widget Function(
-  void Function() goToPreviousTab,
-  void Function() goToNextTab,
-);
-
-/// category header tile builder
-/// allow custom category header tile
-typedef CategoryDayViewHeaderTileBuilder = Function(
-  BoxConstraints constraints,
-  EventCategory category,
 );
 
 /// time label builder
 /// allow custom time label
 typedef TimeLabelBuilder = Widget Function(BuildContext context, DateTime time);
 
-/// sort events in a time range for overflow day view
-typedef OverflowEventsSorter = int Function(DayEvent eventA, DayEvent eventB);